+++ packages/widgetbook_core/lib/src/navigation_tree/widgets/navigation_tree.dart
@@ -59,6 +59,7 @@ class NavigationTreeState extends State<NavigationTree> {
         data: filteredNodes[index],
         selectedNode: selectedNode,
         onNodeSelected: (node) {
+          if (node.path == selectedNode?.path) return;
           setState(() => selectedNode = node);
           widget.onNodeSelected?.call(node.path, node.data);
         },